+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.7-pre
-*
+* Engine template file extension is now optional, defaults to html
 
 ### 0.0.6
 * Fix incorrect URL path and directory path on Windows